requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.4.1
h11==0.12.0
httpcore==0.15.0
httptools==0.6.1
httpx==0.23.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
openai==0.28.0
orjson==3.10.3
pillow==10.3.0
pydantic==2.7.3
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.18.0
python-dotenv==0.21.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.28.1
rfc3986==1.5.0
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
SpeechRecognition==3.8.1
starlette==0.37.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==1.26.18
uvicorn==0.18.2
watchdog==2.2.1
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4